Publish portable symbol packages from project builds
