+++ PowerForge.Tests/DotNetRepositoryReleasePreparationServiceTests.cs
@@ -33,6 +33,7 @@ public sealed class DotNetRepositoryReleasePreparationServiceTests
                 TimeStampServer = "http://timestamp.test",
                 SignAssemblies = false,
                 SignPackages = false,
+                IncludeSymbols = true,
                 Publish = true,
                 PublishFailFast = true,
                 SkipDuplicate = true
@@ -51,6 +52,7 @@ public sealed class DotNetRepositoryReleasePreparationServiceTests
             Assert.Equal(PowerForge.CertificateStoreLocation.LocalMachine, context.Spec.CertificateStore);
             Assert.False(context.Spec.SignAssemblies);
             Assert.False(context.Spec.SignPackages);
+            Assert.True(context.Spec.IncludeSymbols);
             Assert.Equal("Debug", context.Spec.Configuration);
             Assert.True(context.Spec.Publish);
             Assert.True(context.Spec.PublishFailFast);
