+++ Schemas/project.build.schema.json
@@ -120,6 +120,10 @@
       ],
       "description": "Controls how selected projects are packed. PerProject runs a non-incremental build followed by dotnet pack --no-build for each project. MSBuild/Batch requires OutputPath or StagingPath, generates a temporary MSBuild traversal project, and rebuilds then packs selected projects in one invocation; when no package output path is available it falls back to PerProject. Both strategies verify primary managed assemblies under lib/<tfm>, runtimes/<rid>/lib/<tfm>, and tools/<tfm>/any against exact fresh build target directories before signing or publishing. Native runtime assets and metadata-only packages are not assembly-hash targets. The batch target runs Restore;Rebuild;Pack, requires private feed credentials to be available to dotnet/MSBuild restore, stops on the first project failure, and treats the whole failed batch as failed, so already-produced packages from that batch are not signed or published. Projects without a resolved version are reported as failed skipped projects."
     },
+    "IncludeSymbols": {
+      "type": "boolean",
+      "description": "When true, creates portable .snupkg symbol packages and includes them in publish and staged release outputs."
+    },
     "PublishNuget": {
       "type": "boolean"
     },
